requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pydantic-settings
supabase
httpx
python-multipart
openai
sse-starlette